requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-Bcrypt>=1.0.1
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
psycopg2-binary>=2.9  # only needed if using Postgres
Werkzeug>=2.2
